requirements.txt
notion_client==2.4.0
python-dotenv==1.1.1
python-telegram-bot==22.3
